requirements.txt
requests==2.26.0
pytest==6.2.4
allure-python-commons==2.9.45
allure-pytest==2.9.45